refactor(hostd): split the FlowMux UDP relay into its own module
flowmux.rs had grown to 1509 production lines, past the 1500-line
decomposition gate. The per-association UDP datagram relay and the
address codec it frames peers with are one self-contained concern with
their own wire format, so they move to flowmux/udp_relay.rs and the
codec round-trip test moves with them.

Pure move: every relocated line is byte-identical apart from the
pub(super) markers the module boundary needs. No condition, error
message, or signature changed on the guest egress data path. Parent
line count 1509 -> 1317; mvm-hostd test count unchanged at 1901.
